Make mapCustomerGroup and CUSTOMER_GROUP a shared util and type
