+++ src/domain/types.ts
@@ -1,0 +1,7 @@
+export enum CUSTOMER_GROUP {
+  PRIVATE = 'PRIVATE',
+  COMPANY = 'COMPANY',
+  INTERNAL = 'INTERNAL',
+  NON_BILLABLE = 'NON_BILLABLE',
+  OTHER_ORGANIZATION = 'OTHER_ORGANIZATION',
+}
