+++ src/locales/fi.json
@@ -47,6 +47,18 @@
       "SEA_BUOY_MOORING": "Poiju (merellä)",
       "TRAWLER_PLACE": "Traileripaikka"
     },
+    "customerGroups": {
+      "PRIVATE": "Yksityinen",
+      "COMPANY": "Yritys",
+      "INTERNAL": "Sisäinen",
+      "NON_BILLABLE": "Ei laskutettava",
+      "OTHER_ORGANIZATION": "Yhdistys"
+    },
+    "invoicingTypes": {
+      "DIGITAL_INVOICE": "Laskutus sähköisenä",
+      "ONLINE_PAYMENT": "Verkkomaksu",
+      "PAPER_INVOICE": "Laskutus paperisena"
+    },
     "terminology": {
       "width": "Leveys",
       "length": "Pituus",
@@ -184,11 +196,6 @@
       "municipality": "Kotikunta",
       "address": "Jakeluosoite",
       "berths": "Venepaikat",
-      "privateCustomer": "Yksityinen",
-      "companyCustomer": "Yritys",
-      "internalCustomer": "Sisäinen",
-      "nonBillableCustomer": "Ei laskutettava",
-      "otherOrganizationCustomer": "Yhdistys",
       "emptyState": "Rajauksella ei löytynyt asiakastietoja"
     },
     "customerTableTools": {
